backend/requirements.txt
# Sanad Backend - FastAPI
# Pinned versions per EU CRA compliance

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# WebSocket Support
websockets==12.0

# MQTT Client (für IoT-Geräte: NFC-Reader, LED-Controller)
aiomqtt==2.0.0
paho-mqtt==1.6.1

# Push Notifications (FCM)
firebase-admin==6.4.0

# HTTP Client (for external services)
httpx==0.26.0

# Utilities
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
aiosqlite==0.20.0

# Development
black==23.12.1
ruff==0.1.11
